refactor(cli): fix lint and syntax errors in decomposed CLI
